Fix some issues related to hydra_main
This commit provides the following fixes:

- The `version_base` argument is now specified in `hydra.main`
decorator, so no warning message will be raised;
- The `mod` parameter in method `FourDVarNetHydraRunner.test` was not
defined before use. The method has been rewritten to make it clearer;
- If a hydra config key is mandatory and yet unspecified (`???`), an
error message will be raised at a reading attempt. However, when the
configs are converted into a dictionary, the error message will not be
raised, leading sometimes to inconsistent types comparison. The absence
of mandatory hydra config key value is now checked in the `main`.